Fix(be): 호스트가 아닌 유저도 예약을 취소할 수 있도록 변경

+++ backend/grabtable/src/main/java/edu/skku/grabtable/reservation/service/ReservationService.java
@@ -74,8 +74,16 @@ public class ReservationService {
         return ReservationDetailResponse.of(reservation);
     }
 
-    public void deleteByHostId(Long hostId) {
-        Reservation reservation = reservationRepository.findByHostId(hostId)
+    public void cancel(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new BadRequestException(ExceptionCode.NOT_FOUND_USER_ID));
+
+        if (user.getInvitedReservation() != null) {
+            user.clearReservation();
+            return;
+        }
+
+        Reservation reservation = reservationRepository.findByHostId(userId)
                 .orElseThrow(() -> new BadRequestException(ExceptionCode.NO_RESERVATION_USER));
 
         reservationRepository.delete(reservation);
